Do not rely on receive timeout of serial port
With Windows 10 and Arduino MEGA and DUE as ModbusRTU slaves I have experienced that the serial receive timeout is not reliable. In many cases it immediately returns instead of waiting the set timeout time. This leads to errors even tough the slave answered immediately.
Additionally handling the receive timeout in the python script fixed this issue with unreliable driver implementations for me.